+++ main.py
@@ -3,16 +3,18 @@ import os
 from voice import Voice
 import elevenlabs as el
 from dialogue import Dialogue
+from video import Video
+from moviepy.editor import VideoFileClip, AudioFileClip
 
-def GetClonedVoices():
+def GetVoices():
     '''
-    Fetches and returns a user's cloned voices
+    Fetches and returns a user's available voices
     Args: N/A
     Returns:
         [Voice]: Array of Voice objects
     '''
 
-    response = requests.get(el.BASE_URL + "/v2/voices?category=cloned", 
+    response = requests.get(el.BASE_URL + "/v2/voices", 
         headers=el.HEADERS)
     response.raise_for_status() # Throws error if status 4xx or 5xx
     voices = []
@@ -24,12 +26,12 @@ def Main():
     if (not os.getenv('ELEVEN_LABS_API_KEY')):
         return print('API key not found')
     try:
-        voices = GetClonedVoices()
+        voices = GetVoices()
         if (len(voices)):
-            voice1 = voices[0]
-            voice2 = voices[1]
-            d = Dialogue(voice1, voice2, "Yo whats up mr white wanna cook bitch?/Jesse go fucking kill yourself retard")
-            d.CreateAllSpeech()
+            voice1 = voices[43]
+            voice2 = voices[42]
+            d = Dialogue(voice1, voice2, "hello mr white how are you today?/I am doing amazing jesse!")
+            v = Video('new', d)
 
     except requests.exceptions.HTTPError as e:
         # Failed with status [status]: [message]
@@ -37,7 +39,10 @@ def Main():
         f' {e.response.json()['detail']['message']}')
     except NameError as e:
         print(e)
-    except OSError:
+    except OSError as e:
+        print(e)
         print('A system error has occured')
+    except ValueError as e:
+        print(e)
 
 Main()